Issue #1527 Use one raw price loader for all Azure Storage price list loaders
