+++ billing-report-agent/src/main/java/com/epam/pipeline/billingreportagent/service/impl/converter/AbstractAzureStoragePriceListLoader.java
@@ -21,33 +21,21 @@ import com.epam.pipeline.billingreportagent.model.billing.StoragePricing;
 import com.epam.pipeline.billingreportagent.model.pricing.AzurePricingMeter;
 import com.epam.pipeline.billingreportagent.model.pricing.AzurePricingResult;
 import com.epam.pipeline.billingreportagent.service.impl.loader.CloudRegionLoader;
-import com.epam.pipeline.billingreportagent.service.impl.pricing.AzurePricingClient;
 import com.epam.pipeline.entity.region.AbstractCloudRegion;
 import com.epam.pipeline.entity.region.AzureRegion;
 import com.epam.pipeline.entity.region.CloudProvider;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.microsoft.azure.credentials.ApplicationTokenCredentials;
-import lombok.extern.slf4j.Slf4j;
-import okhttp3.OkHttpClient;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.util.Assert;
 import retrofit2.Response;
-import retrofit2.Retrofit;
-import retrofit2.converter.jackson.JacksonConverterFactory;
 
-import java.io.File;
-import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;
@@ -59,17 +47,13 @@ public abstract class AbstractAzureStoragePriceListLoader implements StoragePric
     protected static final String STORAGE_CATEGORY = "Storage";
     protected static final String DATA_STORE_METER_TEMPLATE = "%s Data Stored";
 
-    private static final int READ_TIMEOUT = 30;
-    private static final int CONNECT_TIMEOUT = 30;
-    private static final String API_VERSION = "2016-08-31-preview";
-    private static final String AUTH_TOKEN_TEMPLATE = "Bearer %s";
-    private static final String FILTER_TEMPLATE =
-        "OfferDurableId eq '%s' and Currency eq 'USD' and Locale eq 'en-US' and RegionInfo eq 'US'";
-
     private CloudRegionLoader regionLoader;
+    private AzureRawPriceLoader rawPriceLoader;
 
-    public AbstractAzureStoragePriceListLoader(final CloudRegionLoader regionLoader) {
+    public AbstractAzureStoragePriceListLoader(final CloudRegionLoader regionLoader,
+                                               final AzureRawPriceLoader rawPriceLoader) {
         this.regionLoader = regionLoader;
+        this.rawPriceLoader = rawPriceLoader;
     }
 
     @Override
@@ -86,17 +70,12 @@ public abstract class AbstractAzureStoragePriceListLoader implements StoragePric
             .collect(Collectors.toList());
         Assert.isTrue(CollectionUtils.isNotEmpty(activeAzureRegions), "No Azure regions found in current deployment!");
         assertOffersAndRegionMetersAreUnique(activeAzureRegions);
-        final Map<String, Response<AzurePricingResult>> azureRegionPricing = new HashMap<>();
-        for (final AzureRegion region : activeAzureRegions) {
-            final String offerAndSubscription = getRegionOfferAndSubscription(region);
-            if (!azureRegionPricing.containsKey(offerAndSubscription)) {
-                azureRegionPricing.put(offerAndSubscription, getAzurePricing(region));
-            }
-        }
+        final Map<String, Response<AzurePricingResult>> azureRegionPricing =
+            rawPriceLoader.getRawPricesForRegion(activeAzureRegions);
         return activeAzureRegions.stream()
             .collect(Collectors.toMap(AbstractCloudRegion::getRegionCode,
                 region -> {
-                    final String offerAndSubscription = getRegionOfferAndSubscription(region);
+                    final String offerAndSubscription = rawPriceLoader.getRegionOfferAndSubscription(region);
                     return getStoragePricingForRegion(azureRegionPricing.get(offerAndSubscription),
                                                       region.getMeterRegionName());
                 },
@@ -132,7 +111,7 @@ public abstract class AbstractAzureStoragePriceListLoader implements StoragePric
 
     private void assertOffersAndRegionMetersAreUnique(final List<AzureRegion> activeAzureRegions) {
         final Map<String, Set<String>> regionOfferMapping =
-            mapRegionsOnFunctionResult(activeAzureRegions, this::getRegionOfferAndSubscription);
+            mapRegionsOnFunctionResult(activeAzureRegions, rawPriceLoader::getRegionOfferAndSubscription);
         final Map<String, Set<String>> regionMeterMapping =
             mapRegionsOnFunctionResult(activeAzureRegions, AzureRegion::getMeterRegionName);
         for (final String regionName : regionOfferMapping.keySet()) {
@@ -162,18 +141,6 @@ public abstract class AbstractAzureStoragePriceListLoader implements StoragePric
                                                Function.identity())));
     }
 
-    private ApplicationTokenCredentials getCredentialsFromFile(final String credentialsPath) {
-        try {
-            return ApplicationTokenCredentials.fromFile(new File(credentialsPath));
-        } catch (IOException e) {
-            throw new IllegalArgumentException("Can't access given Azure auth file!");
-        }
-    }
-
-    private String getRegionOfferAndSubscription(final AzureRegion region) {
-        return String.join(",", region.getPriceOfferId(), region.getSubscription());
-    }
-
     private StoragePricing getStoragePricingForRegion(final Response<AzurePricingResult> pricingResponse,
                                                       final String meterRegion) {
         final List<AzurePricingMeter> azurePricingMeters = Optional.ofNullable(pricingResponse.body())
@@ -186,38 +153,5 @@ public abstract class AbstractAzureStoragePriceListLoader implements StoragePric
                                                              meterRegion));
         }
         return storagePricing;
-    }
-
-    private Response<AzurePricingResult> getAzurePricing(final AzureRegion region) throws IOException {
-        final ApplicationTokenCredentials credentials = getCredentialsFromFile(region.getAuthFile());
-        final String resourceManagerEndpoint = credentials.environment().resourceManagerEndpoint();
-        final Response<AzurePricingResult> pricingResponse = buildRetrofitClient(resourceManagerEndpoint)
-            .getPricing(String.format(AUTH_TOKEN_TEMPLATE,
-                                      credentials.getToken(resourceManagerEndpoint)),
-                        region.getSubscription(),
-                        String.format(FILTER_TEMPLATE, region.getPriceOfferId()),
-                        API_VERSION).execute();
-        if (!pricingResponse.isSuccessful()) {
-            throw new IllegalStateException(String.format("Can't load Azure price list for region with id=%s!",
-                                                          region.getId()));
-        }
-        return pricingResponse;
-    }
-
-    private AzurePricingClient buildRetrofitClient(final String azureApiUrl) {
-        final OkHttpClient client = new OkHttpClient.Builder()
-            .followRedirects(true)
-            .readTimeout(READ_TIMEOUT, TimeUnit.SECONDS)
-            .connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
-            .hostnameVerifier((s, sslSession) -> true)
-            .build();
-        final ObjectMapper mapper = new ObjectMapper()
-            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-        final Retrofit retrofit = new Retrofit.Builder()
-            .baseUrl(azureApiUrl)
-            .addConverterFactory(JacksonConverterFactory.create(mapper))
-            .client(client)
-            .build();
-        return retrofit.create(AzurePricingClient.class);
     }
 }
