+++ billing-report-agent/src/main/resources/application.properties
@@ -43,6 +43,7 @@ sync.storage.historical.billing.generation=false
 #sync.storage.s3.disable=true
 #sync.storage.efs.disable=true
 #sync.storage.gs.disable=true
+sync.storage.azure.price.retention.minutes=30
 #sync.storage.azure-blob.disable=true
 sync.storage.azure-blob.category=General Block Blob
 sync.storage.azure-blob.redundancy=LRS
